requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.4.0=pypi_0
aiohttp=3.10.5=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.13.2=pypi_0
anyio=4.5.0=pypi_0
apache-airflow=2.10.1=pypi_0
apache-airflow-providers-common-compat=1.2.0=pypi_0
apache-airflow-providers-common-io=1.4.0=pypi_0
apache-airflow-providers-common-sql=1.16.0=pypi_0
apache-airflow-providers-fab=1.3.0=pypi_0
apache-airflow-providers-ftp=3.11.0=pypi_0
apache-airflow-providers-http=4.13.0=pypi_0
apache-airflow-providers-imap=3.7.0=pypi_0
apache-airflow-providers-smtp=1.8.0=pypi_0
apache-airflow-providers-sqlite=3.9.0=pypi_0
apispec=6.6.1=pypi_0
argcomplete=3.5.0=pypi_0
asgiref=3.8.1=pypi_0
attrs=24.2.0=pypi_0
babel=2.16.0=pypi_0
blinker=1.8.2=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
cachelib=0.9.0=pypi_0
certifi=2024.8.30=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
clickclick=20.10.2=pypi_0
colorama=0.4.6=pypi_0
colorlog=6.8.2=pypi_0
configupdater=3.2=pypi_0
confluent-kafka=2.5.3=pypi_0
connexion=2.14.2=pypi_0
cron-descriptor=1.4.5=pypi_0
croniter=3.0.3=pypi_0
cryptography=43.0.1=pypi_0
deprecated=1.2.14=pypi_0
dill=0.3.8=pypi_0
dnspython=2.6.1=pypi_0
docutils=0.21.2=pypi_0
email-validator=2.2.0=pypi_0
flask=2.2.5=pypi_0
flask-appbuilder=4.5.0=pypi_0
flask-babel=2.0.0=pypi_0
flask-caching=2.3.0=pypi_0
flask-jwt-extended=4.6.0=pypi_0
flask-limiter=3.8.0=pypi_0
flask-login=0.6.3=pypi_0
flask-session=0.5.0=pypi_0
flask-sqlalchemy=2.5.1=pypi_0
flask-wtf=1.2.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.9.0=pypi_0
google-re2=1.1.20240702=pypi_0
googleapis-common-protos=1.65.0=pypi_0
greenlet=3.1.0=pypi_0
grpcio=1.66.1=pypi_0
gunicorn=23.0.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.2=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.4.0=pypi_0
importlib-resources=6.4.5=pypi_0
inflection=0.5.1=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.4=pypi_0
jmespath=1.0.1=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
lazy-object-proxy=1.10.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
limits=3.13.0=pypi_0
linkify-it-py=2.0.3=pypi_0
lockfile=0.12.2=pypi_0
mako=1.3.5=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
marshmallow=3.22.0=pypi_0
marshmallow-oneofschema=3.1.1=pypi_0
marshmallow-sqlalchemy=0.28.2=pypi_0
mdit-py-plugins=0.4.2=pypi_0
mdurl=0.1.2=pypi_0
methodtools=0.4.7=pypi_0
more-itertools=10.5.0=pypi_0
multidict=6.1.0=pypi_0
ncurses=6.4=h6a678d5_0
openssl=3.0.15=h5eee18b_0
opentelemetry-api=1.27.0=pypi_0
opentelemetry-exporter-otlp=1.27.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.27.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.27.0=pypi_0
opentelemetry-exporter-otlp-proto-http=1.27.0=pypi_0
opentelemetry-proto=1.27.0=pypi_0
opentelemetry-sdk=1.27.0=pypi_0
opentelemetry-semantic-conventions=0.48b0=pypi_0
ordered-set=4.1.0=pypi_0
packaging=24.1=pypi_0
pathspec=0.12.1=pypi_0
pendulum=3.0.0=pypi_0
pip=24.2=py311h06a4308_0
pluggy=1.5.0=pypi_0
prison=0.2.1=pypi_0
protobuf=4.25.5=pypi_0
psutil=6.0.0=pypi_0
psycopg2-binary=2.9.9=pypi_0
pycparser=2.22=pypi_0
pygments=2.18.0=pypi_0
pyjwt=2.9.0=pypi_0
python=3.11.9=h955ad1f_0
python-daemon=3.0.1=pypi_0
python-dateutil=2.9.0.post0=pypi_0
python-nvd3=0.16.0=pypi_0
python-slugify=8.0.4=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.32.3=pypi_0
requests-toolbelt=1.0.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rich=13.8.1=pypi_0
rich-argparse=1.5.2=pypi_0
rpds-py=0.20.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=75.1.0=py311h06a4308_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=1.4.54=pypi_0
sqlalchemy-jsonfield=1.0.2=pypi_0
sqlalchemy-utils=0.41.2=pypi_0
sqlite=3.45.3=h5eee18b_0
sqlparse=0.5.1=pypi_0
tabulate=0.9.0=pypi_0
tenacity=9.0.0=pypi_0
termcolor=2.4.0=pypi_0
text-unidecode=1.3=pypi_0
time-machine=2.15.0=pypi_0
tk=8.6.14=h39e8969_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
uc-micro-py=1.0.3=pypi_0
unicodecsv=0.14.1=pypi_0
universal-pathlib=0.2.5=pypi_0
urllib3=2.2.3=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.44.0=py311h06a4308_0
wirerope=0.4.7=pypi_0
wrapt=1.16.0=pypi_0
wtforms=3.1.2=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.11.1=pypi_0
zipp=3.20.2=pypi_0
zlib=1.2.13=h5eee18b_1
